Improved automatic detection of Julia backends in Kernel Tuner and testing
